docs: refactor function renderer
