+++ site/src/components/FunctionRenderer/FunctionRenderer.tsx
@@ -1,0 +1,34 @@
+import React from "react";
+import Heading from "@theme/Heading";
+import Markdown from "markdown-to-jsx";
+import { Function } from "../../doc-types";
+import { renderParameterSignature, renderReturnSignature } from "./helpers";
+
+interface FunctionRendererProps extends Function {
+	children: React.ReactNode;
+}
+
+export const FunctionRenderer: React.FC<FunctionRendererProps> = ({
+	name,
+	description = "No description provided.",
+	parameters = [],
+	returns = [],
+	children,
+}) => {
+	if (description instanceof Array) {
+		description = description.join("\n");
+	}
+
+	return (
+		<section>
+			<Heading as="h3" id={name}>
+				<code>
+					{name}({parameters.length > 0 && renderParameterSignature(parameters)})
+					{returns.length > 0 && `: ${renderReturnSignature(returns)}`}
+				</code>
+			</Heading>
+			<Markdown options={{ forceBlock: true }}>{description}</Markdown>
+			{children}
+		</section>
+	);
+};
