+++ site/src/components/FunctionRenderer/index.ts
@@ -1,0 +1,1 @@
+export { FunctionRenderer as default } from "./FunctionRenderer";
